// src/app/modules/baseinfo/models/security-user.ts
export class SecurityUser
{
    id: number;
    username: string;
    companyEmail: string;
    lastName: string;
    firstName: string;
    persianLastName: string;
    persianFirstName: string;
    gender: number;
    personelCode: string;
    personalEmail: string;
    phone: string;
    isActive: boolean;
    password: string;
    cellPhone: string;
    startJob: Date;
    address: string;
    noMandateTimesheet: string;
}